initial idea
A port of the Plait type-inference utilities to Rhombus, plus a
skeleton of how we can implement Shplait pieces with macros, instead
of (like Plait) having a complicated traverse-the-program loop. The
main strategy is to create a tree of environment frames at compile
time while expanding, and then to process those frames after forcing
the expansion of a modudle body.