more consistent thunk skppin gin lazy mode for value forms
A practical lazy language's implementation will not thunk immediate
values. The Shplait impleemntation previously had one thunk-avoidance
case, which was for immediate identifiers passed to a function ---
intended not as a performance improvement, but to reduce pickiness in
test coverage. For example the `mt_env` argument will not get used in
`interp(parse('1+2'), mt_env)`, so test coverage would report `mt_env`
as unused.

As an example of how `trace` exposes the choice,

```
~lazy
fun f(x):  x
def three: (1 + 2)
// three
trace f: to_string(f(three))
```

the argument to `f` would prints as `#<thunk>` or `3` depending on whether the
`// three` line is uncommented.

Reasonable choices to improve include (1) extending the non-thunking
of arguments to value forms in general; and (2) having a simpler
language, although without a practical coverage assessment. The former
seems more useful and realistic, and I could not resist the minor
extra challenge in the Shplait implementaiton (i.e., to show that it
could be done), so we go with it.

Leaning more on "value" expression checking exposed some problems
there, which are also fixed in this commit.

Related to #3