change to `block`, add implicit `block`s, and add `begin`
An explicit `block` form was intended to help students avoid the
mistake of writing two separate expressions and forgetting to combine
them somehow. But in parallel to Rhombus, `block` is also explained as
the form to support local definitions. The problem is that students
then know about `block` and expect to use it, and they sometimes write
two expressions in `block` where one is ignored.

This commit changes `block` to still allow a mixture of definitions
and expressions, but require that all expressions except the last has
type `Void`, which all but prevents writing a useless expresion whose
result is ignored. Since the main danger of multiple expressions is
thus removed, we go further in making `block` implicit in places like
the body of `fun` or the branch-result positions of `match`, `cond`,
and `if`. Finally, because we eventually need to ignore non-`Void`
results, this commit adds `begin` (matching Moe with state, although
generalized to multiple expressions instead of just two).