fix inference of a polymorphic type for a constructed RHS
In something like `def z = zero()`, where `zero` is a variant of a
polymorphic datatype, the defined name `z` was not given a polymorhpic
type. Instead, a polymorphic type was given only when the number of
fields for the variant happened to match the number of type arguments
for its datatype.